server/requirements.txt
alembic==1.15.2
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloud-init==24.4.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
configparser==7.2.0
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.9
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.7.0
entrypoints==0.3
eventlet==0.39.1
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
google-auth==2.35.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
middleware==1.2.3
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==25.0
pexpect==4.6.0
pipenv==2025.0.2
platformdirs==4.3.8
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
pytest==8.3.5
pytest-flask==1.3.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.1.0
python-engineio==4.12.0
python-http-client==3.3.7
python-socketio==5.13.0
PyYAML==5.3.1
requests==2.32.3
requests-unixsocket==0.2.0
rsa==4.9.1
SecretStorage==2.3.1
sendgrid==6.12.0
service-identity==18.1.0
setuptools==79.0.1
simple-websocket==1.1.0
simplejson==3.16.0
six==1.17.0
sos==4.7.2
SQLAlchemy==2.0.40
ssh-import-id==5.10
starkbank-ecdsa==2.2.0
systemd-python==234
Twisted==18.9.0
typing_extensions==4.13.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
utils==1.0.2
virtualenv==20.31.2
virtualenv-clone==0.3.0
wadllib==1.3.3
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.34.2
wsproto==1.2.0
zipp==1.0.0
zope.interface==4.7.1
